Change collection names from singular to plural in contact and lead controller queries
The collection names in several queries in the contact and lead controllers have been changed from singular to plural. This affects the following files:

- server/controllers/contact/contact.js
- server/controllers/lead/lead.js

In each file, there are multiple instances where the collection name has been updated. For example, in `server/controllers/contact/contact.js`, the collection name `Contact` has been changed to `Contacts` in several `$lookup` stages of aggregation pipelines. Similarly, in `server/controllers/lead/lead.js`, the collection name `Lead` has been changed to `Leads` in several places.

Note that this change was made assuming that the collection names in the database are plural. If this is not the case, the code will need to be adjusted accordingly.